requirements.txt
Django>=4.2.16,<4.3.0
djangorestframework>=3.15.2,<3.16.0
psycopg2-binary>=2.9.5,<3.0.0
Pillow>=11.0.0,<11.1.0
python-dotenv>=1.0.0,<1.1.0
django_simple_captcha>=0.5.18,<0.6.0
requests>=2.32.0,<2.33.0
gunicorn>=22.0.0,<22.1.0
django-cors-headers>=4.3.1,<4.4
django-filter>=24.3,<24.4
user-agents>=2.2.0,<2.3.0
#mysqlclient>=2.2.0,<2.3.0
